docs(components): attachment, Avatar, AvatarGroup, Badge, Button
