+++ src/components/Attachment/__stand__/Attachment.dev.stand.mdx
@@ -13,6 +13,10 @@ import { AttachmentExampleIcon } from './examples/AttachmentExampleIcon/Attachme
 import { AttachmentExampleName } from './examples/AttachmentExampleName/AttachmentExampleName';
 import { AttachmentExampleAs } from './examples/AttachmentExampleAs/AttachmentExampleAs';
 
+```tsx
+import { Attachment } from '@consta/uikit/Attachment';
+```
+
 <MdxMenu>
 
 - [Как использовать](#как-использовать)
@@ -117,7 +121,7 @@ import { AttachmentExampleAs } from './examples/AttachmentExampleAs/AttachmentEx
 ### Иконка
 
 Иконка файла зависит от его расширения, за это отвечает свойство `fileExtension`. Если вы укажете расширение, которого нет в нашей галерее иконок, пользователь увидит иконку с вопросом.
-[Посмотреть иконки для файлов](?path=/story/components-file--gallery)
+[Посмотреть иконки для файлов](##LIBS.STAND|stand:uikit-components-icon-stable)
 
 ```tsx
 <Attachment
@@ -240,7 +244,7 @@ import { AttachmentExampleAs } from './examples/AttachmentExampleAs/AttachmentEx
 Кнопка нужна, чтобы совершать с файлом дополнительные действия (например, удалить его).
 
 `buttonIcon` отвечает за иконку на кнопке, она появится при наведении на компонент.
-Иконку можно выбрать из [библиотеки иконок](?path=/story/components-icons--playground).
+Иконку можно выбрать из [библиотеки иконок](##LIBS.STAND|stand:uikit-components-icon-stable).
 
 `buttonTitle` — всплывающая подсказка, появится при наведении на кнопку.
