+++ src/components/Attachment/__stand__/Attachment.stand.mdx
@@ -1,4 +1,10 @@
-Показывает загрузку файла или уже загруженные файлы.
+import { AttachmentExampleDescription } from './examples/AttachmentExampleDescription/AttachmentExampleDescription';
 
-Можно использовать, когда пользователю нужно загрузить файл — фотографию, скриншот,
+```tsx
+import { Attachment } from '@consta/uikit/Attachment';
+```
+
+Этот компонент можно использовать, когда пользователю нужно загрузить файл — фотографию, скриншот,
 документ или что-то другое. А потом посмотреть, какие документы он загрузил.
+
+<AttachmentExampleDescription />
