+++ src/components/Avatar/__stand__/Avatar.dev.stand.mdx
@@ -3,6 +3,10 @@ import { AvatarExampleImage } from './examples/AvatarExampleImage/AvatarExampleI
 import { AvatarExampleSize } from './examples/AvatarExampleSize/AvatarExampleSize';
 import { AvatarExampleForm } from './examples/AvatarExampleForm/AvatarExampleForm';
 
+```tsx
+import { Avatar } from '@consta/uikit/Avatar';
+```
+
 <MdxMenu>
 
 - [Аватар](#аватар)
