+++ src/components/Avatar/__stand__/Avatar.stand.mdx
@@ -1,5 +1,13 @@
+import { AvatarExampleImage } from './examples/AvatarExampleImage/AvatarExampleImage';
+
+```tsx
+import { Avatar } from '@consta/uikit/Avatar';
+```
+
 Аватар — графическое изображение пользователя, фотография или рисунок.
 
 Он нужен, чтобы быстро узнавать пользователя. Например, в чате или в ветке комментариев.
 А ещё — чтобы пользователь мог узнать сам себя. Например, проверить, что вошёл в
 аккаунт под нужным логином.
+
+<AvatarExampleImage />
