+++ src/components/Badge/__stand__/Badge.stand.mdx
@@ -1,6 +1,9 @@
-import { MdxMenu } from '@consta/stand';
 import { BadgeExampleUse } from '../__stories__/examples/BadgeExampleUse/BadgeExampleUse';
 import { BadgeExampleStatus } from '../__stories__/examples/BadgeExampleStatus/BadgeExampleStatus';
+
+```tsx
+import { Badge } from '@consta/uikit/Badge';
+```
 
 Бейджик показывает статус объекта в каком-то контексте.
 
@@ -14,23 +17,6 @@ import { BadgeExampleStatus } from '../__stories__/examples/BadgeExampleStatus/B
 Бейджик часто помогает понять, что сейчас происходит с объектом, и подсказывает, что делать дальше.
 
 <BadgeExampleStatus />
-
-<MdxMenu>
-
-- [Как использовать](#как-использовать)
-- [Статус и содержимое](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:статус-и-содержимое)
-  - [Бейджик с текстом](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:бейджик-с-текстом)
-  - [Бейджик без текста](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:бейджик-без-текста)
-  - [Бейджик с иконкой](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:бейджик-с-иконкой)
-- [Внешний вид](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:внешний-вид)
-  - [Типы бейджиков](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:типы-бейджиков)
-  - [Размер](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:размер)
-  - [Форма](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:форма)
-- [Импорт](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:импорт)
-- [Список свойств](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:свойства)
-- [Пример использования](##LIBS.STAND.DEV|stand:uikit-components-badge-stable|hash:пример)
-
-</MdxMenu>
 
 ## Как использовать
