+++ src/components/Button/__stand__/examples/ButtonExampleAdaptive/ButtonExampleAdaptive.tsx
